moved graphql logic to separate client
